Added some animations
The navbar has been changed to show which page the user is currently at with an underline. 
When the user hovers over the navigation items, there is a change of colour with a line going from left to right.
In the contact page, animation has been added to the button where hovering over the button will initiate the animation from left to right.